fully test util
